+++ worker_test.go
@@ -280,11 +280,12 @@ type testMetricExporter interface {
 	Metric() (*metricpb.Metric, error)
 }
 
-func exportMetricAndFlush(t testing.TB, exp testMetricExporter) WorkerMetrics {
+func exportMetricAndFlush(t testing.TB, exp testMetricExporter, scope metricpb.Scope) WorkerMetrics {
 	w := NewWorker(1, nil, logrus.New(), nil)
 	m, err := exp.Metric()
 	assert.NoErrorf(t, err, "exporting the metric '%s' shouldn't have failed",
 		exp.GetName())
+	m.Scope = scope
 
 	assert.NoError(t, w.ImportMetricGRPC(m), "importing a metric shouldn't "+
 		"have failed")
@@ -297,23 +298,49 @@ func TestWorkerImportMetricGRPC(t *testing.T) {
 		h := samplers.NewHist("test.histo", nil)
 		h.Sample(1.0, 1.0)
 
-		assert.Len(t, exportMetricAndFlush(t, h).histograms, 1,
+		wms := exportMetricAndFlush(t, h, metricpb.Scope_MIXED)
+		assert.Len(t, wms.histograms, 1,
 			"The number of flushed histograms is not correct")
+	})
+	t.Run("global histogram", func(t *testing.T) {
+		t.Parallel()
+		h := samplers.NewHist("test.histo", nil)
+		h.Sample(1.0, 1.0)
+
+		wms := exportMetricAndFlush(t, h, metricpb.Scope_GLOBAL)
+		assert.Len(t, wms.globalHistograms, 1,
+			"The number of flushed global histograms is not correct")
+	})
+	t.Run("local histogram should fail", func(t *testing.T) {
+		t.Parallel()
+		h := samplers.NewHist("test.histo", nil)
+		h.Sample(1.0, 1.0)
+
+		w := NewWorker(1, nil, logrus.New(), nil)
+		m, err := h.Metric()
+		assert.NoErrorf(t, err, "exporting the metric '%s' shouldn't have failed",
+			h.GetName())
+
+		m.Scope = metricpb.Scope_LOCAL
+
+		assert.Error(t, w.ImportMetricGRPC(m), "importing a local-only metric "+
+			"should have failed")
 	})
 	t.Run("gauge", func(t *testing.T) {
 		t.Parallel()
 		g := samplers.NewGauge("test.gauge", nil)
 		g.Sample(2.0, 1.0)
 
-		assert.Len(t, exportMetricAndFlush(t, g).globalGauges, 1,
-			"The number of flushed gauges is not correct")
+		wms := exportMetricAndFlush(t, g, metricpb.Scope_GLOBAL)
+		assert.Len(t, wms.globalGauges, 1, "The number of flushed gauges is not correct")
 	})
 	t.Run("counter", func(t *testing.T) {
 		t.Parallel()
 		c := samplers.NewCounter("test.counter", nil)
 		c.Sample(2.0, 1.0)
 
-		assert.Len(t, exportMetricAndFlush(t, c).globalCounters, 1,
+		wms := exportMetricAndFlush(t, c, metricpb.Scope_GLOBAL)
+		assert.Len(t, wms.globalCounters, 1,
 			"The number of flushed counters is not correct")
 	})
 	t.Run("timer", func(t *testing.T) {
@@ -324,11 +351,26 @@ func TestWorkerImportMetricGRPC(t *testing.T) {
 
 		m, err := h.Metric()
 		assert.NoErrorf(t, err, "exporting the histogram shouldn't have failed")
+		m.Scope = metricpb.Scope_MIXED
 		m.Type = metricpb.Type_Timer
+
+		assert.NoError(t, w.ImportMetricGRPC(m), "importing a metric shouldn't "+
+			"have failed")
+		assert.Len(t, w.Flush().timers, 1, "The number of flushed timers is not correct")
+	})
+	t.Run("global timer", func(t *testing.T) {
+		w := NewWorker(1, nil, logrus.New(), nil)
+		h := samplers.NewHist("test.timer.global", nil)
+		h.Sample(1.0, 1.0)
+
+		m, err := h.Metric()
+		assert.NoErrorf(t, err, "exporting the histogram shouldn't have failed")
+		m.Type = metricpb.Type_Timer
+		m.Scope = metricpb.Scope_GLOBAL
 
 		assert.NoError(t, w.ImportMetricGRPC(m), "importing a timer shouldn't "+
 			"have failed")
-		assert.Len(t, w.Flush().timers, 1, "The number of flushed "+
+		assert.Len(t, w.Flush().globalTimers, 1, "The number of flushed "+
 			"timers is not correct")
 	})
 	t.Run("set", func(t *testing.T) {
@@ -336,8 +378,8 @@ func TestWorkerImportMetricGRPC(t *testing.T) {
 		s := samplers.NewSet("test.set", nil)
 		s.Sample("value", 1.0)
 
-		assert.Len(t, exportMetricAndFlush(t, s).sets, 1,
-			"The number of flushed sets is not correct")
+		wms := exportMetricAndFlush(t, s, metricpb.Scope_MIXED)
+		assert.Len(t, wms.sets, 1, "The number of flushed sets is not correct")
 	})
 }